[REFACTOR] Controller와 Service 코드 주석 처리
